requirements.txt
pandas
google-cloud-bigquery
google-auth
ftfy
python-dotenv
db-dtypes
